Rename tower enums to enums
