+++ src/models/Tower.ts
@@ -7,7 +7,7 @@ import {
     Unique,
     OneToOne,
 } from "typeorm"
-import { TowerType, TowerKingdom, TowerLevel } from "../definitions/TowerEnums"
+import { TowerType, TowerKingdom, TowerLevel } from "../definitions/enums"
 import { MainStats } from "./MainStats"
 import { BarracksStats } from "./BarracksStats"
 import { AttackStats } from "./AttackStats"
